Fix code style for weka resources

+++ spec/weka/classifiers/meta/one_class_classifier_spec.rb
@@ -1,34 +1,39 @@
 require 'spec_helper'
 
 describe Weka::Classifiers::Meta::OneClassClassifier do
+  it { is_expected.to be_a Java::WekaClassifiersMeta::OneClassClassifier }
 
-  it { should be_a Java::WekaClassifiersMeta::OneClassClassifier }
+  let(:classifier_type) { 'Meta::OneClassClassifier' }
 
   before do
     @config = test_config
-    @w_options = "-W weka.classifiers.meta.Bagging -- -W weka.classifiers.trees.RandomForest -- -I 100"
-    options = "-tcl #{ Wikipedia::VandalismDetection::Instances::VANDALISM } #{@w_options}"
+    classifier_options = '-W weka.classifiers.trees.RandomForest -- -I 100'
+    @w_options = "-W weka.classifiers.meta.Bagging -- #{classifier_options}"
+    vandalism = Wikipedia::VandalismDetection::Instances::VANDALISM
+    options = "-tcl #{vandalism} #{@w_options}"
 
-    @config.instance_variable_set :@classifier_type, 'Meta::OneClassClassifier'
-    @config.instance_variable_set :@classifier_options, options
-    @config.instance_variable_set :@cross_validation_fold, 2
+    @config.instance_variable_set(:@classifier_type, classifier_type)
+    @config.instance_variable_set(:@classifier_options, options)
+    @config.instance_variable_set(:@cross_validation_fold, 2)
 
     use_configuration(@config)
 
-    # add more test instances because instances number must higher than cross validation fold
-    instances = Wikipedia::VandalismDetection::TrainingDataset.instances.to_a.map(&:values)
+    # add more test instances because instances number must be higher than
+    # cross validation fold:
+    data = Wikipedia::VandalismDetection::TrainingDataset.instances.to_m.to_a
     dataset = Wikipedia::VandalismDetection::Instances.empty
 
     2.times do
-      instances.each do |row|
+      data.each do |row|
         values = row[0..-2]
-        class_value = Wikipedia::VandalismDetection::Instances::CLASSES[rand((0..1))]
+        index = rand((0..1))
+        class_value = Wikipedia::VandalismDetection::Instances::CLASSES[index]
+
         dataset.add_instance([*values, class_value])
       end
     end
 
     Wikipedia::VandalismDetection::TrainingDataset.stub(instances: dataset)
-    puts Wikipedia::VandalismDetection::TrainingDataset.instances
   end
 
   after do
@@ -37,42 +42,39 @@ describe Weka::Classifiers::Meta::OneClassClassifier do
 
     if File.exist?(arff_file)
       File.delete(arff_file)
-      FileUtils.rm_r(File.dirname arff_file)
+      directory = File.dirname(arff_file)
+      FileUtils.rm_r(directory)
     end
 
-    if Dir.exist?(build_dir)
-      FileUtils.rm_r(build_dir)
-    end
+    FileUtils.rm_r(build_dir) if Dir.exist?(build_dir)
   end
 
-  it "can be used to classify vandalism" do
-    expect {
-      classifier = Wikipedia::VandalismDetection::Classifier.new
-      features = [1.0, 2.0, 55.0]
-      confidence = classifier.classify(features)
-      puts "vandalism confidence: #{confidence}}"
-    }.not_to raise_error
+  it 'can be used to classify vandalism' do
+    classifier = Wikipedia::VandalismDetection::Classifier.new
+    features = [1.0, 2.0, 55.0]
+
+    expect { classifier.classify(features) }.not_to raise_error
   end
 
-  it "can be used to classify vandalism using regulars" do
-    options = "-tcl #{ Wikipedia::VandalismDetection::Instances::REGULAR } #{@w_options}"
+  it 'can be used to classify vandalism using regulars' do
+    regular = Wikipedia::VandalismDetection::Instances::REGULAR
+    options = "-tcl #{regular} #{@w_options}"
 
-    @config.instance_variable_set :@classifier_type, 'Meta::OneClassClassifier'
-    @config.instance_variable_set :@classifier_options, options
+    @config.instance_variable_set(:@classifier_type, classifier_type)
+    @config.instance_variable_set(:@classifier_options, options)
 
     use_configuration(@config)
 
-    expect {
-      classifier = Wikipedia::VandalismDetection::Classifier.new
-      features = [1.0, 2.0, 8.0]
-      confidence = classifier.classify(features)
-      puts "regular confidence: #{confidence}}"
-    }.not_to raise_error
+    classifier = Wikipedia::VandalismDetection::Classifier.new
+    features = [1.0, 2.0, 8.0]
+
+    expect { classifier.classify(features) }.not_to raise_error
   end
 
-  describe "#type" do
-    it "returns the classifier's type name" do
-      expect(Weka::Classifiers::Meta::OneClassClassifier.type).to eq 'Meta::OneClassClassifier'
+  describe '#type' do
+    it 'returns the classifier’s type name' do
+      expect(Weka::Classifiers::Meta::OneClassClassifier.type)
+        .to eq 'Meta::OneClassClassifier'
     end
   end
 end
